Add files via upload
Doodle project for testing out the Keras Tensorflow API for python. The project uses a Long-Short-Term-Memory Neural Network to predict future daily prices of GBP/EUR. The network is trained using an 80-10-10 split of training, validation and testing data and currently performs at a loss value of 1.17*10e-4 after 100 epochs (which is completely unnecessary). In the future I will be experimenting with this ML project to grow my knowledge of time series forecasting.
FYI a decent chunk of this code has been reused from various sources online that I can no longer find to cite.